[REF] Main Content 클래스 내부 컴포넌트 분리
